Extract CalledMethodProcessor from NodeScopeResolver
