+++ src/Testing/TypeInferenceTestCase.php
@@ -65,7 +65,6 @@ abstract class TypeInferenceTestCase extends PHPStanTestCase
 	{
 		$container = self::getContainer();
 		$reflectionProvider = self::createReflectionProvider();
-		$typeSpecifier = $container->getService('typeSpecifier');
 
 		$enableFnsr = getenv('PHPSTAN_FNSR');
 		$className = NodeScopeResolver::class;
@@ -80,10 +79,8 @@ abstract class TypeInferenceTestCase extends PHPStanTestCase
 			$container->getExtensionsCollection(FunctionParameterOutTypeExtension::class),
 			$container->getExtensionsCollection(MethodParameterOutTypeExtension::class),
 			$container->getExtensionsCollection(StaticMethodParameterOutTypeExtension::class),
-			self::getParser(),
 			$container->getByType(FileTypeMapper::class),
 			$container->getByType(PhpDocInheritanceResolver::class),
-			$container->getByType(FileHelper::class),
 			$container->getExtensionsCollection(ReadWritePropertiesExtension::class),
 			$container->getExtensionsCollection(FunctionParameterClosureThisExtension::class),
 			$container->getExtensionsCollection(MethodParameterClosureThisExtension::class),
@@ -92,7 +89,6 @@ abstract class TypeInferenceTestCase extends PHPStanTestCase
 			$container->getExtensionsCollection(MethodParameterClosureTypeExtension::class),
 			$container->getExtensionsCollection(StaticMethodParameterClosureTypeExtension::class),
 			$container->getExtensionsCollection(PerFileAnalysisResettable::class),
-			self::createScopeFactory($reflectionProvider, $typeSpecifier),
 			$container->getParameter('polluteScopeWithLoopInitialAssignments'),
 			$container->getParameter('polluteScopeWithAlwaysIterableForeach'),
 			$container->getParameter('exceptions')['implicitThrows'],
