+++ tests/PHPStan/Analyser/AnalyserTest.php
@@ -820,10 +820,8 @@ class AnalyserTest extends PHPStanTestCase
 			$container->getExtensionsCollection(FunctionParameterOutTypeExtension::class),
 			$container->getExtensionsCollection(MethodParameterOutTypeExtension::class),
 			$container->getExtensionsCollection(StaticMethodParameterOutTypeExtension::class),
-			$this->getParser(),
 			$fileTypeMapper,
 			$phpDocInheritanceResolver,
-			$fileHelper,
 			$container->getExtensionsCollection(ReadWritePropertiesExtension::class),
 			$container->getExtensionsCollection(FunctionParameterClosureThisExtension::class),
 			$container->getExtensionsCollection(MethodParameterClosureThisExtension::class),
@@ -832,7 +830,6 @@ class AnalyserTest extends PHPStanTestCase
 			$container->getExtensionsCollection(MethodParameterClosureTypeExtension::class),
 			$container->getExtensionsCollection(StaticMethodParameterClosureTypeExtension::class),
 			$container->getExtensionsCollection(PerFileAnalysisResettable::class),
-			self::createScopeFactory($reflectionProvider, $typeSpecifier),
 			false,
 			true,
 			true,
